requirements-sheets.txt
asgiref==3.5.2
black==22.10.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
EZSheets==2022.10.10
google-api-core==2.10.2
google-api-python-client==1.7.9
google-auth==2.13.0
google-auth-httplib2==0.0.3
google-auth-oauthlib
googleapis-common-protos==1.56.4
gsheets==0.6.1
gspread==5.6.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.2.2
pathspec==0.10.1
platformdirs==2.5.2
protobuf==4.21.8
psycopg2==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.5
PyJWT==2.5.0
pyparsing==3.0.9
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==3.0.1
urllib3==1.26.12
